OnPoint-014 Fix waiting for recycler view function [GM]
The wait does not work properly anymore given the larger sizes of the
recyclerviews, making them always displayed, even if they contain
nothing. Instead we need to check that it has a descendant with
some text we know it will appear in the list. It's not the best solution
as we assume something on the content we get from Moodle, but there
is not ******* way to make this work properly because Espresso
just does not support the most basic operations like checking if
a recyclerview has elements. We cannot even do this programmatically
as we are unable to retrieve the view by its ID with the activity
test rule, given that the activity has changed after switching tabs
(no idea how, it's still the same one, just a different fragment)
which causes null pointer exceptions.