OnPoint-023 Fix double-entry in the displayed list [SF,GM]
Assignments were inserted in the DB once, but displayed twice in the
list due to a redundant addition to the list and notification to the
recyclerview adapter.

Also added a few more toast messages to notify the user about the
download of assignments happening and of how many new ones were found.